requirements.txt
flask
openai
pillow
requests
gunicorn
werkzeug
python-dotenv